finish new backend base structure
