+++ api/src/middlewares/authentication.handler.js
@@ -1,0 +1,6 @@
+import passport from 'passport';
+
+const authenticateJwt = passport.authenticate('jwt', { session: false });
+const authenticateLocal = passport.authenticate('local', { session: false });
+
+export { authenticateJwt, authenticateLocal };
