+++ api/src/middlewares/error.handler.js
@@ -1,6 +1,7 @@
-// middleware for errors with Boom library:
+// Middlewares for handling errors:
+
+// Handler for Boom errors:
 const boomErrorHandler = (err, req, res, next) => {
-  console.log(`BOOM ERROR: ${err}`);
   if (err.isBoom) {
     const { output } = err;
     res.status(output.statusCode).json(output.payload);
@@ -8,6 +9,7 @@ const boomErrorHandler = (err, req, res, next) => {
   next(err);
 };
 
+// Handler for generic errors:
 const genericErrorHandler = (err, req, res, next) => {
   const status = err.status || 500;
   const message = err.message || 'Internal Server Error';
