+++ api/src/middlewares/passport.handler.js
@@ -1,0 +1,11 @@
+// External packages:
+import passport from 'passport';
+
+// Internal packages:
+import LocalStrategy from '../utils/authenticationStrategies/local.strategy.js';
+import JwtStrategy from '../utils/authenticationStrategies/jwt.strategy.js';
+
+passport.use(LocalStrategy);
+passport.use(JwtStrategy);
+
+export default passport;
