+++ api/src/services/user/user.service.js
@@ -1,18 +1,24 @@
+// External packages:
 import bcrypt from 'bcrypt';
 import Boom from '@hapi/boom';
-import { UserModel } from '../../Db/Schema/user.js';
 
+// Internal packages:
+import { UserModel } from '../../db/schemas/user.js';
+
+// UserService class:
 class UserService {
   constructor() {}
 
-  // service to create user:
+  // Service to create user:
   async createUser(data) {
     // Check if username exists:
+
     const existingUser = await this.getUser(data.username);
 
     if (existingUser) {
       throw Boom.badRequest('Username already exists');
     }
+
     // hash password:
     const hash = await bcrypt.hash(data.password, 10);
 
@@ -28,13 +34,13 @@ class UserService {
     };
   }
 
-  // service to get user:
+  // Service to get user:
   async getUser(username) {
     const existingUser = await UserModel.findOne({ username });
     return existingUser;
   }
 
-  // service to get user by id:
+  // Service to get user by id:
   async getUserById(id) {
     const user = await UserModel.findById(id);
     return {
