--- api/src/utils/authentication/index.js
@@ -1,8 +1,0 @@
-import passport from 'passport';
-import LocalStrategy from './strategies/local.strategy.js';
-import JwtStrategy from './strategies/jwt.strategy.js';
-
-passport.use(LocalStrategy);
-passport.use(JwtStrategy);
-
-export default passport;
