+++ client/src/Components/Experience/Background/Background courses/Background_Courses.jsx
@@ -46,7 +46,6 @@ const CourseItem = ({ course }) => {
       });
       return;
     }
-    console.log(user);
     dispatch(getFile(user.token, key));
   };
